requirements.txt
hikari
hikari-lightbulb
pymongo
requests
python-dotenv
bs4
mal-api
AnilistPython
